1차 마감 controller 및 view 코딩컨벤션 점검

+++ src/main/java/com/fms/smartbutler/service/UsersService.java
@@ -98,6 +98,10 @@ public class UsersService implements UserDetailsService {
 	}
 	
 	public void delete(UsersDTO usersDTO) {
+		if(usersDTO.getStatus() == 2) {
+			return;
+		}
+		
 		Users users = modelMapper.map(usersDTO, Users.class);
 		
 		usersRepository.delete(users);
